% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.8.30.R
\name{interpolate}
\alias{interpolate}
\title{interpolate}
\usage{
interpolate(stand.bp, stand.time, sample)
}
\description{
This function interpolates the size of each peak in the query given a ladder of know size
}
\examples{
interpolate()
}
\keyword{lm}
