% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.8.30.R
\name{plot.fsa.all}
\alias{plot.fsa.all}
\title{Plot fsa all Function}
\usage{
\method{plot}{fsa.all}(
  file_path,
  ladder,
  cutoff = 250,
  revise.cutoff = T,
  sd = 3,
  channel.ladder = 4,
  channel.query = 1
)
}
\arguments{
\item{revise.cutoff}{Would you like to change the cutoff for finding peaks? Defaults to TRUE.}
}
\description{
This function allows you to visualize the chromatogram. Plotsall channels
}
\examples{
plot.fsa.all()
}
\keyword{fsa}
