% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.10.18.R, R/Functions.8.30.R,
%   R/Functions.9.01.R
\name{plot.fsa}
\alias{plot.fsa}
\title{Plot fsa Function}
\usage{
\method{plot}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  revise.cutoff = T,
  sd = 3,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)

\method{plot}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  revise.cutoff = T,
  sd = 3,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)

\method{plot}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  revise.cutoff = T,
  sd = 3,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)
}
\arguments{
\item{revise.cutoff}{Would you like to change the cutoff for finding peaks? Defaults to TRUE.}
}
\description{
This function allows you to visualize the chromatogram.Only plots ladder and query channel

This function allows you to visualize the chromatogram.Only plots ladder and query channel

This function allows you to visualize the chromatogram.Only plots ladder and query channel
}
\examples{
plot.fsa()
plot.fsa()
plot.fsa()
}
\keyword{fsa}
