% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.10.18.R, R/Functions.8.30.R,
%   R/Functions.9.01.R
\name{sum.fsa}
\alias{sum.fsa}
\title{Summerize fsa Function}
\usage{
\method{sum}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  sd = 3,
  revise.cutoff = T,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)

\method{sum}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  sd = 3,
  revise.cutoff = T,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)

\method{sum}{fsa}(
  file_path,
  ladder = rox.ladder,
  cutoff = 250,
  sd = 3,
  revise.cutoff = T,
  only.up = T,
  channel.ladder = 4,
  channel.query = 1
)
}
\arguments{
\item{revise.cutoff}{Would you like to change the cutoff for finding peaks? Defaults to TRUE.}
}
\description{
This function allows you to visualize the chromatogram.Only plots ladder and query channel
the cutoff applies to all channels
Revise cutoff only applies to the ladder channel
In the ladder channel fragment sizes are assigned to peaks from the largest to smallest fragments of DNA.
if revise.cutoff = TRUE, then all peaks in the ladder channel are called,
then the median and sd are calculated from the tallest peaks.
If the ladder has 16 fragments than the 16 largest peaks will be used for this calculation.
anything under the number of specified standard deviations from the median is removed when the
revised cutoff is implemented.

This function allows you to visualize the chromatogram.Only plots ladder and query channel
the cutoff applies to all channels
Revise cutoff only applies to the ladder channel
In the ladder channel fragment sizes are assigned to peaks from the largest to smallest fragments of DNA.
if revise.cutoff = TRUE, then all peaks in the ladder channel are called,
then the median and sd are calculated from the tallest peaks.
If the ladder has 16 fragments than the 16 largest peaks will be used for this calculation.
anything under the number of specified standard deviations from the median is removed when the
revised cutoff is implemented.

This function allows you to visualize the chromatogram.Only plots ladder and query channel
the cutoff applies to all channels
Revise cutoff only applies to the ladder channel
In the ladder channel fragment sizes are assigned to peaks from the largest to smallest fragments of DNA.
if revise.cutoff = TRUE, then all peaks in the ladder channel are called,
then the median and sd are calculated from the tallest peaks.
If the ladder has 16 fragments than the 16 largest peaks will be used for this calculation.
anything under the number of specified standard deviations from the median is removed when the
revised cutoff is implemented.
}
\examples{
sum.fsa()
sum.fsa()
sum.fsa()
}
\keyword{fsa}
\keyword{summary}
