% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.8.30.R
\name{extrapolate}
\alias{extrapolate}
\title{extrapolate}
\usage{
extrapolate(stand.bp, stand.time, sample)
}
\description{
This function extrapolate the size of each peak in the query given a ladder of know size
}
\examples{
interpolate()
}
\keyword{lm}
