% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.10.18.R, R/Functions.8.30.R,
%   R/Functions.9.01.R
\name{extrapolate}
\alias{extrapolate}
\title{extrapolate}
\usage{
extrapolate(stand.bp, stand.time, sample)

extrapolate(stand.bp, stand.time, sample)

extrapolate(stand.bp, stand.time, sample)
}
\arguments{
\item{revise.cutoff}{default = TRUE}
}
\description{
This function extrapolate the size of each peak in the query given a ladder of know size

This function extrapolate the size of each peak in the query given a ladder of know size

This function extrapolate the size of each peak in the query given a ladder of know size
}
\examples{
interpolate()
interpolate()
interpolate()
}
\keyword{lm}
