% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.8.30.R
\name{filter.bp}
\alias{filter.bp}
\title{filter.bp}
\usage{
filter.bp(res.bp, bp.min = 200, bp.max = 600)
}
\arguments{
\item{revise.cutoff}{Defaults bp.min = 200,  bp.max = 600.}
}
\description{
This function removes DNA fragments outside a given size range
}
\examples{
filter.bp()
}
\keyword{fsa}
