% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.8.30.R
\name{pairwise.dist}
\alias{pairwise.dist}
\title{Pairwise frag summary}
\usage{
pairwise.dist(str.1 = database[[1]], str.2, window = 2.5, method = "bc")
}
\arguments{
\item{window}{the largest difference in fragment size (bp) between two
fragement summaries at which peaks/fragments are considered an observation
of the same fragment default is 2.5bp.}

\item{method}{Default distacne is Bray-Curtis,"jd" Jaccard}
}
\description{
This function allows you calculate the Bray Curtis (BC) or Jaccard distance between two
fragment summaries
}
\examples{
pairwise.dist()
}
\keyword{BC}
\keyword{Bray-Curtis}
\keyword{fsa}
\keyword{summary}
